requirements.txt
rhasspy-hermes==0.1.3
attrs==19.3.0
paho-mqtt==1.5.0
